==> 00_Codigos.ipynb <==
## Codigos xx cepas.
Este programa sirve para generar un archivo .txt que contenga las lineas de codigo que necesito para correr las 170 cepas del paper [Genomic Survey of Bordetella pertussis](https://wwwnc.cdc.gov/eid/article/25/4/18-0812_article) en el programa ISCompare.py

# en primer lugar necesitamos generar dos tuplas, una que contenga los nombres de las cepas y otras que contenga
# los accesion number

# abrimos el archivo 'Genomic Survey of Bordetella pertussis.xls'
import pandas as pd

df = pd.read_excel('Genomic_Survey_of_Bordetella_pertussis.xlsx')
df

# quiero ordenar la columna por ano de aislamiento

df = df.sort_values('Year', ascending=True, ignore_index=True)
df

# reseteo el indice
df.reset_index(drop=True, inplace=True)
df.head(50)

# necesito obtener los datos de la columna 'ID' como una lista o una tupla.
# Lo mismo con los datos de la columna 'Accession no.'
cepas = df['ID'].to_list()
acc_no = df['Accession no.'].to_list()


# ademas necestio generar una DF que solo tenga la columna de las 'ID'
id_df = df[['ID']]
id_df.to_csv('Nombres_Cepas.csv', index=False)

# con las dos listas que genere antes ahora tengo que crear un diccionario. Para eso uso la funcion zip()
# mis_datos = dict(zip(claves, valores))

dic = dict(zip(cepas,acc_no))
print(dic)


# ahora creo el archivo txt sobre el cual voy a escribir todos los codigos que necesito
file = open('codigos_170_cepas.txt', 'w')

i=0
for cepas, acc_no  in dic.items():
    i= i+1
    file.write('\n{} --- {} ---\n'.format(i, cepas))
    file.write('ISCompare.py -i IS481_M28220.1.fasta -Q {} -R BX470248.1 -e nambrosis@hotmail.com -o {} -c\n'.format(acc_no, cepas))

file.close()
