==> 01_Contador de ISs.ipynb <==
### Objetivo
Uno de los resultados que arroja el programa ISCompare es el numero de ISs encontradas en la cepa de Referencia y en las cepas analizada.

El objetivo de este bloque de codigo es extraer esa informacion del archivo 'stats.txt' y generar una DataFrame y un archivo .csv que condense esta informacion.

'''
nota II: cuidado que en en los alrchivos stats.txt siempre hay 2 lineas que tienen info sobre el numero de ISs que 
se encontraron en la cepa query. Habria que analizar con mauri cual es la que hay que usar. Yo creo que es la linea 0 y
es la que use para este programa
'''

import pandas as pd
pd.set_option('display.max_rows', 175)
import re



strains_list = [] 
print(f'\nCEPAS >>> {strains_list}')
print(f'\nNumero de cepas: {len(strains_list)}')




ISs_num = {}

############################################################################################################
def get_ISs_from_ref(strains_list): 
    
    try: #con este try el programa no sea cae si no encuentra la carpeta que busca. Al final avisa que carpeta no encontro.
            with open(f'./{strains_list[0]}/' + 'stats.txt', 'r+') as f:
                file = f.readlines()
                f.close()
                lines_list = []
                
                for line in file:                    
                    nombre = line.replace("\n", "")
                    lines_list.append(nombre)
        
                number = r"\d{1,3}"
                ref_num =int(re.search(number, lines_list[11]).group())
                strain = 'Tohama'
                ISs_num[strain] = ref_num
    except Exception:            
            print(f'The folder ::{strain}:: was not found.')
                
    return ISs_num

############################################################################################################

def get_IS(lines_list):
    for line in lines_list:
        number = r"\d{1,3}"
        querry_num =int(re.search(number, lines_list[0]).group()) 
        return querry_num
############################################################################################################

def get_ISs_from_querys(strains_list):
    
    for strain in strains_list:
        
        try:
            
            with open(f'./{strain}/' + 'stats.txt', 'r+') as f:
                
                file = f.readlines()
                f.close()

                lines_list = []
                for line in file:
                    
                    nombre = line.replace("\n", "")
                    lines_list.append(nombre)

                ISs_num[strain] = get_IS(file) 

        except Exception:
            
            print(f'The folder ::{strain}:: was not found.')

    return ISs_num
############################################################################################################
get_ISs_from_ref(strains_list)
get_ISs_from_querys(strains_list)
print(ISs_num)

df = pd.DataFrame.from_dict(ISs_num,orient='index', columns=["ISs number"])
display(df.sort_values(by=["ISs number"]))

# notar que Tohama es siempre la cepa que menos ISs tiene


df.to_csv('ISs_number.csv')
print('File has been saved! ')

# un paso interesante seria graficar en forma automatica el nombre el numero de ISs

---
# Data Visualization
---

strains = []
ISs = []
for key, value in ISs_num.items():
    strains.append(key)
    ISs.append(value)


print(f'Strains >>> {strains}')
print(f'ISs >>> {ISs}')


# libraries
import numpy as np
import matplotlib.pyplot as plt

# Make dataset
height = ISs 
bars = strains 
y_pos = np.arange(len(bars))
 
# Define a specific color for Tohama bar
colors = ['#ff9696' if i == 'Tohama' else '#73c3ff' for i in bars]
    
# Create horizontal bars
plt.figure(figsize=(9, 7))

plt.barh(y_pos, height, color=colors)
 
# Create names on the y-axis
plt.yticks(y_pos, bars)
plt.xlabel("Number of IS's", fontsize=12)
plt.ylabel('Strains', fontsize=12)


# Add a line indicating Tohama ISs number
plt.axvline(x=234, color='Black', linestyle='dashed')
 
# Show graphic
plt.show()
