==> 02_Concatenate.ipynb <==
---
### Analisis de Secuencias de Insercion IS 481 en aislamientos clinicos de B. pertussis Vs Tohama
---

### Bibliotecas

import pandas as pd
pd.set_option('display.max_columns', 85)
pd.set_option('display.max_rows', 100)

### Concatenacion de resultados

# Este bloque de codigo permite obtener una lista con los nombres de las 170 cepas 
# del paper "Genomic Survey of Bordetella pertussis"



df = pd.read_csv('Nombres_Cepas.csv')

strains_list = df['ID'].tolist()

print(strains_list)


# Este bloque de codigo me sirve para concatenar los archivos 'FinalResults.csv' que salen de correr el 
# ISCompare con las cepas
# el programa tiene que ser ejecutado en la misma carpeta en donde se encuentran las carpetas con los nombres de las cepas.




final = pd.DataFrame()
for strain in strains_list:
    try: #con este try el programa no sea cae si no encuentra la carpeta que busca. Al final avisa que carpeta no encontro.
        df = pd.read_csv(f'./{strain}/' + 'FinalResults.csv') 
        final = pd.concat([final,df], ignore_index=True) 

    except Exception:
        print(f'The folder ::{strain}:: was not found.')
   



    

### Contador de cepas

numero_cepas = final['Query.ID1'].nunique() #recordar que la funcion .nunique() no cuenta los NaN
print(f'Se ha concatenado los resultados de >>> {numero_cepas} <<< cepas')

print(final.shape)
final
# estaria bueno eliminar la columna que se llama 'Unnamed: 0'

final.drop(columns='Unnamed: 0', inplace=True) #esta columna se debe generar como parte del proceso de concatenar

print(final.shape)
final

final.to_csv('Final_Results_concat.csv', index=False)

# NOTA: recordar que el archivo Final_Results_concat.csv contiene los dos tipos de DLIS
#                     > is only present in Ref
#                     > is only present in Query


## Siguiente... 
El siguiente paso es utilzar el programa Merge para agragar la informacion de cada cepa que obtuvimos del paper Genomic Survey...
