==> 03_Merge.ipynb <==
# Programa Merge
Este programa permite combinar el archivo el archivo Final_Results_concat.csv de las DLIS de todas las cepas con el el archivo Genomic_Survey_of_Bordetella_pertussis.xlsx en donde estan las caracteristicas de las cepas. De esta manera obtenemos un .csv que combina las posiciones de las ISs con las caracteristicas de las cepas (ej: año de aislamiento, genotipo, etc.)

## Import

import pandas as pd
pd.set_option('display.max_columns', 45)

---

strains_df = pd.read_excel('Genomic_Survey_of_Bordetella_pertussis.xlsx')

strains_df

strains_df.head(3) # hay algunas columnas que no tiene sentido que la arrastremos para el analisis
#                   asi que las voy a eliminar

strains_df.drop(columns=['State','EPS', 'MLVA', 'wgsMLVA', 'PFGE', 'Depth', 'Reference'], inplace=True)
# Eliminamos todas esas columnas

print(strains_df.shape)
strains_df

# Le voy a cambiar el nombre a la columna 'Accession no.' para que luego conicida con el nombre de la columna de la otra df
# y podemas hacer el merge

strains_df.rename(inplace=True, columns={'Accession no.':'Query_ID'})

strains_df.head(0) #confirmamos que modificamos el nombre de la columna

ISs_df = pd.read_csv('Final_Results_concat.csv') # abrimos el segundo archivo y creamos una nueva df

print(ISs_df.shape)
ISs_df

# le cambiamos el nombre a la columna 'Query.ID1' para que coincida con el nombre de la columna de la otra df
ISs_df.rename(inplace=True, columns={'Query.ID1':'Query_ID'})

ISs_df.head(0)

# En la columna 'Query_ID' hay celdas que estan vacias que me van a molestar para hacer el MERGE en los pasos siguientes.
# Las tengo que eliminar

ISs_df = ISs_df[ISs_df['Query_ID'].isna() == False] # con esta linea sobreescribimos la DF ISs_df con aquellas filas
# que no tienen valores NaN

print(ISs_df.shape)
ISs_df

# Este bloque sirve para elimir el '.1' de todas las celdas de la columna 'Query_ID'. Este paso es necesario
# para poder hacer el merge con la otra DF

ISs_df['Query_ID'] = ISs_df['Query_ID'].str.slice(0,-2)


print(ISs_df.shape)
ISs_df

---
## Merge

print(f'strains_df >>> {strains_df.shape}')
print(f'ISs_df >>> {ISs_df.shape}')
display(strains_df.head(), ISs_df.head() )

# con esta linea hacemos el MERGE!
merge_df = pd.merge(ISs_df, strains_df, left_on='Query_ID', right_on='Query_ID', how='left')

print(f'merge_df >>> {merge_df.shape}')
merge_df

# Gurdamos el archivo

merge_df.to_csv('ISs_Strains.csv', index=False)


---

## Siguiente...
El siguiente paso es analizar los resultados para lo cual se puede pasar al programa DLIS_analysis que tiene escritos codigos para hacer este analisis.
