==> 04_Data_Cleaning.ipynb <==
Este programa permite limpiar las columnas 'REF_Flank1.locus_tag' y 'REF_Flank2.locus_tag' de forma tal de que en cada celda solo se informe UN locus bajo el formato BP1234. Las funciones definidas en este bloques de codigo han sido pensadas para modificar las celdas dando como resultado el locus que se encuentra mas cerca de la IS.

Las diferentes formas en que se presentaban los datos en las celdas de las columnas 'REF_Flank1.locus_tag' y 'REF_Flank2.locus_tag' han sido caracterizadas con nombres que van desde type_1 a type_6 para 'REF_Flank1.locus_tag' y type_1 a type_8 para 'REF_Flank2.locus_tag'. Mas abajo se describe el tipo de dato para cada clasificacion.


Luego se adicionarion lineas de codigo para combinar las columnas 'REF_Flank1.locus_tag' y 'REF_Flank2.locus_tag' de forma tal de conseguir UNA columna que contenga ambos datos.




import pandas as pd
pd.set_option('display.max_columns', 45)
pd.set_option('display.max_rows', 350)

df = pd.read_csv('ISs_Strains.csv')
df.columns

### #1 Data Cleanning columna 'REF_Flank1.locus_tag' 

# df = pd.read_csv('ISs_Strains.csv')
for i in df['REF_Flank1.locus_tag'].unique():
    print(f'\n{i}')

"""
Los resultados que siguen a continuacion son casos especiales que deberiamos limpiar.
IMPORTANTE: notar que cada tipo de dato que queremos limpiar tiene una len() particular, por
lo tanto podemos elegir ese criterio para definir las funciones que vamos a usar para limpiar cada tipo de dato.

type_1: 'BP1047 | BP1048' 

type_2: '-', '- | -', '- | - | -'

type_3: 'BP0123 | BP0124 | -'

type_4: 'BP3053 | BP3054 | BP3055'

type_5: 'BP3508 | -' o 'BP0692 | - | -'

type_6: 'BP2524 | BP2524A'  finalemente decidi que iba a dejar la letra A porque asi esta anotado en Tohama

type_7: 'BP0101B' finalmente no lo considere como un tipo de dato a limpiar porque en el genoma de Tohama esta anotado asi

"""

# type_1: ej 'BP1047 | BP1048'
df = pd.read_csv('ISs_Strains.csv')

for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 15:
        print(i)
    else: pass

# x = 'BP1047 | BP1048'
# x.split(' ')[2]


# type_2: '-', '- | -', '- | - | -'
df = pd.read_csv('ISs_Strains.csv')

for i in df['REF_Flank1.locus_tag'].unique():
    if i == '-' or i == '- | -' or i == '- | - | -':
        print(i)
    else: pass

# type_3: ej 'BP0123 | BP0124 | -'
df = pd.read_csv('ISs_Strains.csv')

for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 19:
        print(i)
    else: pass



# type_4: ej 'BP3053 | BP3054 | BP3055'
df = pd.read_csv('ISs_Strains.csv')

for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 24:
        print(i)
    else: pass

# # type_5: ej 'BP3508 | -' or 'BP0692 | - | -'

df = pd.read_csv('ISs_Strains.csv')

for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 10 or len(i) == 14:
        print(i)
    else: pass

# len('BP0692 | - | -')


# type_6: ej 'BP2524 | BP2524A'

df = pd.read_csv('ISs_Strains.csv')

for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 16:
        print(i)
        
    else: pass
    
# x = 'BP2029 | BP2029A'    

# x.split(' ')[2]

---
## Data Cleanning Functions-Flank1-

############################################

# type_1: ej 'BP1047 | BP1048'>>> for 7<x<16: split(' ') y quedarse con el el elemento [2] 
def clean_type_1(x):    
    if len(x) == 15:
        return x.split(' ')[2]
    else: 
        return x

#######################################################
def clean_type_2(x):
    if x == '-' or x == '- | -' or x == '- | - | -':
        return 'NaN'
        
    else:
        return x
    
#######################################################
# type_3: ej 'BP0123 | BP0124 | -'
def clean_type_3(x):
    if len(x) == 19:
        return x.split(' ')[2]
    else: return x
        
########################################################
# type_4: ej 'BP3053 | BP3054 | BP3055'
def clean_type_4(x):
    if len(x) == 24:
        return x.split(' ')[4]
    else: return x

#######################################################
# type_5: ej 'BP3508 | -'
def clean_type_5(x):
    if len(x) == 10 or len(x) == 14:
        return x.split(' ')[0]
    else: return x

#########################################################
# type_6: ej 'BP2524 | BP2524A'
def clean_type_6(x):
    if len(x) == 16:
        return x.split(' ')[2]
    else: return x
    
##########################################################    



# ejecutando estas lineas se produce la limpieza de las celdas.
df['REF_Flank1.locus_tag'] = df['REF_Flank1.locus_tag'].map(clean_type_1)
df['REF_Flank1.locus_tag'] = df['REF_Flank1.locus_tag'].map(clean_type_2)
df['REF_Flank1.locus_tag'] = df['REF_Flank1.locus_tag'].map(clean_type_3)
df['REF_Flank1.locus_tag'] = df['REF_Flank1.locus_tag'].map(clean_type_4)
df['REF_Flank1.locus_tag'] = df['REF_Flank1.locus_tag'].map(clean_type_5)
df['REF_Flank1.locus_tag'] = df['REF_Flank1.locus_tag'].map(clean_type_6) 
   


# nota:
# me gustaria agregar un contador para saber cuantas veces se corrigieron las celdas para cada uno de los casos,
#pero todavia no me doy cuenta como hacerlo

---
---
---

### Chequeamos que las celdas de 'REF_Flank1.locus_tag' hayan sido corregidas

# Type_1
for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 15:
        print(i)
    else: pass

# type_3
for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 19:
        print(i)
    else: pass 

# type_4
for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 24:
        print(i)
    else: pass

# type_5
for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 10 or len(i) == 14:
        print(i)
    else: pass

# type_6
for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 16:
        print(i)
    else: pass

### Resultado final de Data Cleaning de la columna 'REF_Flank1.locus_tag'

for i in df['REF_Flank1.locus_tag'].unique():
    print(f'\n{i}')
    
# nota: si la data cleanning funciono correctamente cada linea de resultados deberia mostrar SOLO UN locus.

print(f'df >>> {df.shape}')
df

---
---
---


### #2 Data Cleanning 'REF_Flank2.locus_tag' 


for i in df['REF_Flank2.locus_tag'].unique():
    print(f'\n{i}')

"""
Los resultados que siguen a continuacion son casos especiales que deberiamos limpiar.
IMPORTANTE: notar que cada tipo de dato que queremos limpiar tiene una len() particular, por
lo tanto podemos elegir
ese criterio para definir las funciones que vamos a usar para limpiar cada tipo de dato.

type_1: 'BP1047 | BP1048'>>> for 7<x<16: split(' ') y quedarse con el el elemento [3] 

type_2: '-', '- | -', '- | - | -'

type_3: 'BP0123 | BP0124 | -'

type_4: 'BP3053 | BP3054 | BP3055'

type_5: 'BP3508 | -' o 'BP0692 | - | -'

type_6: 'BP2524 | BP2524A'  finalemente decidi que iba a dejar la letra A porque asi esta anotado en Tohama

type_7: 'BP0101B' finalmente no lo considere como un typo de dato a limpiar porque en el genoma de Tohama esta anotado asi

"""

### Chequeamos la presencia de datos a corregir en las celdas de la columna "'REF_Flank2.locus_tag'"

# type_1

for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 15:
        print(i)
    else: pass

# x = 'BP1047 | BP1048'
# x.split(' ')[2]


# type_2: '-', '- | -', '- | - | -'

for i in df['REF_Flank2.locus_tag'].unique():
    if i == '-' or i == '- | -' or i == '- | - | -':
        print(i)
    else: pass

# type_3

for i in df['REF_Flank1.locus_tag'].unique():
    if len(i) == 19:
        print(i)
    else: pass



# type_4

for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 24:
        print(i)
    else: pass

# type_5: ej 'BP3508 | -' or 'BP0692 | - | -'

for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 10 or len(i) == 14:
        print(i)
    else: pass

# len('BP0692 | - | -')


# type_6: ej 'BP2524 | BP2524A'

for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 16:
        print(i)
        
    else: pass
    
# x = 'BP2029 | BP2029A'    

# x.split(' ')[2]

# type_7: ej 'BP1134 | - | - | -'
# este caso no lo tenia del otro flanco

for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 18:
        print(i)
        
    else: pass
# len('BP1134 | - | - | -')

# type_8: ej 'BP1332A | BP1333 | BP1334'

for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 25:
        print(i)
        
    else: pass
# len('BP1332A | BP1333 | BP1334')

---
## Data Cleanning Functions-Flank2-

# df = pd.read_csv('ISs_Strains.csv')
# df
############################################

# type_1: ej 'BP1047 | BP1048'>>> for 7<x<16: split(' ') y quedarse con el el elemento [2] 
def clean_type_1(x):    
    if len(x) == 15:
        return x.split(' ')[0]
    else: 
        return x

#######################################################
def clean_type_2(x):
    if x == '-' or x == '- | -' or x == '- | - | -':
        return 'NaN'
        
    else:
        return x
    
#######################################################
# type_3: ej 'BP0123 | BP0124 | -'
def clean_type_3(x):
    if len(x) == 19:
        return x.split(' ')[0]
    else: return x
        
########################################################
# type_4: ej 'BP3053 | BP3054 | BP3055'
def clean_type_4(x):
    if len(x) == 24:
        return x.split(' ')[0]
    else: return x

#######################################################
# type_5: ej 'BP3508 | -'
def clean_type_5(x):
    if len(x) == 10 or len(x) == 14:
        return x.split(' ')[0]
    else: return x

#########################################################
# type_6: ej 'BP2524 | BP2524A'
def clean_type_6(x):
    if len(x) == 16:
        return x.split(' ')[0]
    else: return x
    
##########################################################  
# type_7: ej 'BP1134 | - | - | -' 
def clean_type_7(x):
    if len(x) == 18:
        return x.split(' ')[0]
    else: return x

##########################################################  
# type_8: ej 'BP1332A | BP1333 | BP1334' 
def clean_type_8(x):
    if len(x) == 25:
        return x.split(' ')[0]
    else: return x


# ejecutando estas lineas se produce la limpieza de las celdas.
df['REF_Flank2.locus_tag'] = df['REF_Flank2.locus_tag'].map(clean_type_1)
df['REF_Flank2.locus_tag'] = df['REF_Flank2.locus_tag'].map(clean_type_2)
df['REF_Flank2.locus_tag'] = df['REF_Flank2.locus_tag'].map(clean_type_3)
df['REF_Flank2.locus_tag'] = df['REF_Flank2.locus_tag'].map(clean_type_4)
df['REF_Flank2.locus_tag'] = df['REF_Flank2.locus_tag'].map(clean_type_5)
df['REF_Flank2.locus_tag'] = df['REF_Flank2.locus_tag'].map(clean_type_6)
df['REF_Flank2.locus_tag'] = df['REF_Flank2.locus_tag'].map(clean_type_7)
df['REF_Flank2.locus_tag'] = df['REF_Flank2.locus_tag'].map(clean_type_8)

---
---
---

### Chequeamos que las celdas de 'REF_Flank2.locus_tag' hayan sido corregidas

# Type_1
for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 15:
        print(i)
    else: pass

# type_3
for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 19:
        print(i)
    else: pass 

# type_4


for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 24:
        print(i)
    else: pass

# type_5


for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 10 or len(i) == 14:
        print(i)
    else: pass

# type_6


for i in df['REF_Flank2.locus_tag'].unique():
    if len(i) == 16:
        print(i)
    else: pass

for i in df['REF_Flank2.locus_tag'].unique():
    print(f'\n{i}')

print(f'df >>> {df.shape}')
df

---
---
---

## Combine REF.Product_Flank1 y REF.Product_Flank2
Vamos a combinar las celdas de las columnas REF.Product_Flank1 y REF.Product_Flank2 de manera tal que al analizar los datos no tengamos que hacer dos analisis por separado y podamos enfocarnos solo en una columna



df['Insertion_site(REF)'] = df['REF_Flank1.locus_tag'] + '-' + df['REF_Flank2.locus_tag']

for i in df['Insertion_site(REF)'].unique():
    print(f'\n{i}')

print(f'df >>> {df.shape}')
print(df["Insertion_site(REF)"])

--- 
# Psrtial Results

print(f'df >>> {df.shape}')
df

--- 
## Save partial results as file.CSV

file = 'ISs_Strains_clean.csv'
df.to_csv(file, index=False)
print(f'File "{file}" has been saved')



---

# A continuacion vamos a agregar una columna que tenga en la que las celdas contengan el numero de cepas que se analizaron
# el anio en que se analizo la cepa en cuestion.

import pickle

with open('str_by_year.pickle', 'rb') as handle:
    b = pickle.load(handle)
    
b

years = []
strains = []
for key, value in b.items():
    years.append(key)
    strains.append(value)

print(years)  
print(strains)

data = {'Year': years, 'Total_strains': strains}
df_years = pd.DataFrame.from_dict(data=data)
df_years

# una vez creado esta DataFrame puedo mergear las dos DataFrames utilzando la columna Year para emparejarlas

print(f'df_years >>> {df_years.shape}')
print(f'df >>> {df.shape}')
display(df_years.head(), df.head())

# con esta linea hacemos el MERGE!
merge_df = pd.merge(df, df_years, left_on='Year', right_on='Year', how='left')

print(f'merge_df >>> {merge_df.shape}')
merge_df

file = 'ISs_Strains_clean.csv'
merge_df.to_csv(file, index=False)
print(f'File "{file}" has been saved')
