==> 05_DLIS_analysis.ipynb <==
# Differentially Located IS's (DLIS)
En este programa analizamos las DLIS encontradas utilizando del [IS-Compare](https://github.com/maurijlozano/ISCompare).
Cepa de Referencia: Tohama///Cepas Query: las 170 cepas del paper 'Genomic Survey of Bordetellla pertussis'.

Este programa se alimenta del archivos 'ISs_Strains_clean.cvs'

---
## Import

import pandas as pd
pd.set_option('display.max_columns', 45)
pd.set_option('display.max_rows', 350)

df = pd.read_csv('ISs_Strains_clean.csv')

print(df.shape)
df

---
## Contador de cepas
Las lineas de codigo que estan a continuacion nos permiten contar cuantas cepas estamos analizando

numero_cepas = df['ID'].nunique()

print(f' Numero de cepas analizadas: >>>>> {numero_cepas}')

---
## DLIS
En el bloque que sigue a continuacion seleccionamos solo los hits que arrojaron DLIS como resultados.
Notar que hay otros hits que se sugiere revisar en forma manual que NO forman parte de este analisis.
nota: ver archivo "Verify manually_analysis"

# Nos quedamos solo con las filas que tienen 'Observation == 1.DLIS'
df = df[df['Observations']=='1. DLIS']

print(df.shape)
df

# Analizamos que tipo de datos estan aojados en cada columna
df.dtypes 

# Para poder trabajar con la funcion .groupby() necesitamos que la columna Year sea un object y no int64
# Por lo tanto modificamos el formato de esa columna
# df['Year'] = df['Year'].astype(object)


# Confirmamos la modificacion del tipo de dato
# df.dtypes

---
## DLIS por cepa por year

# La tabla que se muestra a continuacion muestra el numero de DLIS encontradas por el IS-Compare para cada cepa, agrupadas
# segun el anio en que fueron aisladas

df1 = df.groupby(['Year','ID',]).count()[['Observations']]
df1

# Similar a la anterior, pero ahora tambien agrupamos segun sea que la IS solo esta en Tohama (ref Strain) o solo esta
# en el aislamiento

# IMPORTANTE
# Lo interesante de esta observacion es que pareciera que hay 3 Is's que estaban en Tohama y 
# no estan en los aislamientos clinicos
# utiizamos el SnapGene para confirmar si Tohama tiene una IS481 en esas posiciones
df2 = df.groupby(['Year','ID','Description']).count()[['Observations']]
df2

# las dos tablas de arriba hay que buscar la forma de graficarlas!!!
# convendria hacer 2 tablas la primera para ISs que estan en Tohama y no est'an en las cepas y la segunda al reves

---
## IS's only present in Tohama.
En este bloque analizamos aquellas IS's que estan en Tohama pero NO se encuentran en los aislamientos clinicos (post 2000)

# con este bloque de codigo generamos la df_ref que solo tiene las dlis que ESTAN en Tohama y NO ESTAN en los aislamientos
df_ref = df[df['Description']=='IS only present in ref Strain.']
print(f'df_ref >>> {df_ref.shape}')
df_ref

# Analizamos las posiciones en las que estan las ISs que ESTABAN en Tohama y ya NO ESTAN en los aislamientos
df_ref01 = df_ref.groupby(['Insertion_site(REF)']).count()[['Observations']].sort_values('Observations', ascending = False)
print(f'df_ref >>> {df_ref01.shape}')
df_ref01

# este resultado sugiere que NO hay muchas ISs que Tohama haya perdido, pero las pocas que hay (n=3) parecen haberse
# perdido en casi todos los aislamientos clinicmos analizados
# ojo que el hit que cayo entre BP3508-3511 podria ser en realidad una IS que al salirse activo el gen 3509
# hay que inspeccinar el genoma para intentar descifrar de que caso se trata.
# hay que revisar el hit NaN-BP3409!!

---
## IS's only present in  query strains.
En este bloque analizamos aquellas IS's que estan en los aislamientos (post 2000) pero NO se encuentran en Tohama 

# con este bloque de codigo generamos la df_query que solo tiene las DLIS que estan en los aislamientos 
# y NO estan en Tohama
df_query = df[df['Description']=='IS only present in query Strain.']
print(f'df_query >>> {df_query.shape}')
df_query

# Analizamos al lado de que genes estan las ISs que estan los aislamientos y NO estan en Tohama.
df_query01 = df_query.groupby(['Insertion_site(REF)','Year', 'Total_strains']).count()[['Observations']].sort_values(['Insertion_site(REF)','Year'], ascending = [True,False])


display(df_query01)

'''
analizando esta tabla a simple vista se aprecia que solo la DLIS que cayo en la prn (BP) 1054 esta proporcionalmente mas que
en los anios en los que hubo pico de casos'''

## Genes INTERRUMPIDOS por IS's
En este bloque buscamos IS's que cayeron en el medio de una secuencia codificante es decir que podrian haber interrumpido 
el gen y evitar su transcripcion. Simil a lo que sucede con algunas cepas prn menos.



df_query03 = df_query[df_query['REF_Flank1.locus_tag']==df_query['REF_Flank2.locus_tag']]

print(df_query03.shape)
df_query03

df_query04 = df_query03.groupby(['REF_Flank1.locus_tag']).count()[['Observations']].sort_values('Observations', ascending = False)

df_query04

# recordar que 1054 es la pertactina
# El gen Bp 2611 esta interrumpido en casi todos (143/153!!) los aislamientos clinicos post 2000 (en USA)
#que es la Bp2611? Podria ser un regulador transcripcional!!

# hay 8 cepas en las que BP2611 no esta interrumpida...cuales son? De cuando son?

### Hasta aca llegamos... como seguimos...
