==> 07-Strains_by_state .ipynb <==
El objetivo de este grupo de bloques de código es analizar de qué estados fueron aisladas las cepas que de USA que estamos analizando.

import pandas as pd

file_name = 'Genomic_Survey_of_Bordetella_pertussis.xlsx' # el archivo tiene que estar en la misma carpeta en que se corre 
                                                          # el programa
df = pd.read_excel(file_name)
df

us_state_abbrev = {
    'Alabama': 'AL',
    'Alaska': 'AK',
    'American Samoa': 'AS',
    'Arizona': 'AZ',
    'Arkansas': 'AR',
    'California': 'CA',
    'Colorado': 'CO',
    'Connecticut': 'CT',
    'Delaware': 'DE',
    'District of Columbia': 'DC',
    'Florida': 'FL',
    'Georgia': 'GA',
    'Guam': 'GU',
    'Hawaii': 'HI',
    'Idaho': 'ID',
    'Illinois': 'IL',
    'Indiana': 'IN',
    'Iowa': 'IA',
    'Kansas': 'KS',
    'Kentucky': 'KY',
    'Louisiana': 'LA',
    'Maine': 'ME',
    'Maryland': 'MD',
    'Massachusetts': 'MA',
    'Michigan': 'MI',
    'Minnesota': 'MN',
    'Mississippi': 'MS',
    'Missouri': 'MO',
    'Montana': 'MT',
    'Nebraska': 'NE',
    'Nevada': 'NV',
    'New Hampshire': 'NH',
    'New Jersey': 'NJ',
    'New Mexico': 'NM',
    'New York': 'NY',
    'North Carolina': 'NC',
    'North Dakota': 'ND',
    'Northern Mariana Islands':'MP',
    'Ohio': 'OH',
    'Oklahoma': 'OK',
    'Oregon': 'OR',
    'Pennsylvania': 'PA',
    'Puerto Rico': 'PR',
    'Rhode Island': 'RI',
    'South Carolina': 'SC',
    'South Dakota': 'SD',
    'Tennessee': 'TN',
    'Texas': 'TX',
    'Utah': 'UT',
    'Vermont': 'VT',
    'Virgin Islands': 'VI',
    'Virginia': 'VA',
    'Washington': 'WA',
    'West Virginia': 'WV',
    'Wisconsin': 'WI',
    'Wyoming': 'WY'
}
# este diccionario esta sacado de:https://gist.github.com/rogerallen/1583593

# este bloque nos permite invertir el orden de las key y values del diccionario us_state_abbrev y volver a guardar el nuevo
# diccionario en la misma variable llamada us_state_abbrev
print(f'ANTES >>>\n {us_state_abbrev}')
us_state_abbrev = {v: k for k, v in us_state_abbrev.items()}
print(f'\nAHORA >>>\n {us_state_abbrev}')


# este bloque nos permite cambiar las abreviaturas de la columna States por los nombres completos de los estados
df["State"] = df["State"].map(us_state_abbrev)
df

# obtenemos el primer Heat map con el numero de aislamientos por estado y por anio
states_heatmap = (df
                  .groupby('State')
                  .Year
                  .value_counts()
                  .unstack()
                  .fillna(0)
)


display(states_heatmap)
# este heat map no esta normalizado segun el numero de aislamientos de ese anio. Para ver el analisis normalizado 
# ir hacia el siguiente Heatmap que ademas esta graficado. 

## Heat map

import seaborn as sns
import matplotlib.pyplot as plt

#Create the Figure
fig1, ax1 = plt.subplots(figsize=(15,20))


# Define Data and set parameters
sns.heatmap(states_heatmap,
           annot=True,
           linewidths=0.5,
           linecolor='black',
           cbar=True,
           square=False,
           cmap='BuPu',
            )


# Setting title and labels
plt.title('Number of Isolates by Year\n', fontsize=25)
plt.xlabel('Years', fontsize=18)
plt.ylabel('State', fontsize=18)
plt.yticks(rotation = 0, fontsize=12, fontweight='bold')




# save figure
plt.savefig('Isolates by year-numbers-.png', dpi=300, format='png', bbox_inches='tight')
print('Figure has been saved!!')

# Ahora vamos a normalizar el numero de aislamiento en funcion del numero total de aislamientos obtenidos ese mismo anio.

import pickle
with open('str_by_year.pickle', 'rb') as handle:
    str_by_year = pickle.load(handle)
    
str_by_year

years = list(range(2000,2013+1)) #recordar que la funcion range no incluye al ultimo valor del par

for year in years:
    states_heatmap[year] = states_heatmap[year]/str_by_year[year] 
    
states_heatmap

#con este boque confirmamos que la suma de las fracciones de cada estado da 1. Si no da uno entonces hay un error.
states_heatmap.sum()

## Graficamos el Heat map

import seaborn as sns
import matplotlib.pyplot as plt

#Create the Figure
fig2, ax2 = plt.subplots(figsize=(15,20))


# Define Data and set parameters
sns.heatmap(states_heatmap,
           annot=True,
           linewidths=0.5,
           linecolor='black',
           cbar=True,
           square=False,
           cmap='BuPu',
            )


# Setting title and labels
plt.title('Proportion of Isolates by Year\n', fontsize=25)
plt.xlabel('Years', fontsize=18)
plt.ylabel('State', fontsize=18)
plt.yticks(rotation = 0, fontsize=12, fontweight='bold')




# save figure
plt.savefig('Isolates by year-proportion-.png', dpi=300, format='png', bbox_inches='tight')
print('Figure has been saved!!')
